add checker script

+++ Contest Template/gen.cpp
@@ -12,9 +12,7 @@ const ll LINF = 1e18;
 
 mt19937 rng(chrono::steady_clock::now().time_since_epoch().count());
 
-int rand(int l, int r) {
-    return uniform_int_distribution(l, r)(rng);
-}
+ll rand(ll l, ll r) {return uniform_int_distribution(l, r)(rng);}
 
 int main() {
     ios_base::sync_with_stdio(0);
